pyproject.toml
[project]
name = "langfuse-workshop"
version = "0.1.0"
requires-python = "==3.14.*"
dependencies = [
    "openai>=1.0.0",
    "langfuse>=3.0.0",
    "python-dotenv>=1.0.0",
    "rich>=13.0.0",
    "gradio>=4.0.0",
]
